requirements.txt
langchain
langchain-community
langchain-groq
langchain-text-splitters
sentence-transformers
faiss-cpu
pymupdf
streamlit
python-dotenv
